Add black magic

+++ src/main/java/snownee/snow/ItemSnowLayer.java
@@ -68,7 +68,7 @@ public class ItemSnowLayer extends ItemSnow
             float f = playerIn.rotationPitch;
             float f1 = playerIn.rotationYaw;
             double d0 = playerIn.posX;
-            double d1 = playerIn.posY + (double) playerIn.getEyeHeight();
+            double d1 = playerIn.posY + playerIn.getEyeHeight();
             double d2 = playerIn.posZ;
             Vec3d vec3d = new Vec3d(d0, d1, d2);
             float f2 = MathHelper.cos(-f1 * 0.017453292F - (float) Math.PI);
@@ -78,7 +78,7 @@ public class ItemSnowLayer extends ItemSnow
             float f6 = f3 * f4;
             float f7 = f2 * f4;
             double d3 = playerIn.getEntityAttribute(EntityPlayer.REACH_DISTANCE).getAttributeValue();
-            Vec3d vec3d1 = vec3d.add((double) f6 * d3, (double) f5 * d3, (double) f7 * d3);
+            Vec3d vec3d1 = vec3d.add(f6 * d3, f5 * d3, f7 * d3);
             RayTraceResult result = worldIn.rayTraceBlocks(vec3d, vec3d1, false, false, false);
             if (result != null && result.typeOfHit == RayTraceResult.Type.BLOCK && result.sideHit != EnumFacing.DOWN)
             {
